+++ src/Config/Version/MissingVersionException.php
@@ -1,0 +1,19 @@
+<?php
+declare(strict_types=1);
+
+namespace Smile\GdprDump\Config\Version;
+
+use Exception;
+use Throwable;
+
+class MissingVersionException extends Exception
+{
+    /**
+     * @param string $message
+     * @param Throwable $previous
+     */
+    public function __construct(string $message, Throwable $previous = null)
+    {
+        parent::__construct($message, 0, $previous);
+    }
+}