+++ tests/unit/Config/Version/VersionConditionTest.php
@@ -20,22 +20,9 @@ class VersionConditionTest extends TestCase
     }
 
     /**
-     * Test the "match" method.
-     */
-    public function testConditionMatch()
-    {
-        $condition = new VersionCondition(('<2.3.0'));
-
-        $this->assertTrue($condition->match('2.1'));
-        $this->assertTrue($condition->match('2.1.0-alpha'));
-        $this->assertTrue($condition->match('2.1.0'));
-        $this->assertFalse($condition->match('2.3.0'));
-    }
-
-    /**
      * Test if an exception is thrown when the condition syntax is invalid.
      *
-     * @expectedException \InvalidArgumentException
+     * @expectedException \Smile\GdprDump\Config\Version\InvalidVersionException
      */
     public function testInvalidConditionSyntax()
     {
@@ -45,7 +32,7 @@ class VersionConditionTest extends TestCase
     /**
      * Test if an exception is thrown when the condition does not contain at least 3 characters.
      *
-     * @expectedException \InvalidArgumentException
+     * @expectedException \Smile\GdprDump\Config\Version\InvalidVersionException
      */
     public function testErrorWithLessThanThreeCharacters()
     {