+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## WIP
+
+- Move the config version parsing to a VersionLoader implementation
+
 ## [1.0.0-beta13] - 2019-10-31
 [1.0.0-beta13]: https://git.smile.fr/dirtech/gdpr-dump/compare/1.0.0-beta12...1.0.0-beta13